requirements.txt
fake-bpy-module-2.82==20200514
numpy==1.19.4
OpenEXR==1.3.2
scipy==1.5.4
typing==3.7.4.3
